+++ AFKScape/Assets/Scripts/Storage/Bank.cs
@@ -1,5 +1,6 @@
 ﻿using System.Collections;
 using System.Collections.Generic;
+using System.Numerics;
 using UnityEngine;
 
 public class Bank : Storage
@@ -9,12 +10,12 @@ public class Bank : Storage
         totalSlots = 5000;
     }
 
-    public override void RaiseItemAddedEvent(long id, int amount, int amounDiff)
+    public override void RaiseItemAddedEvent(long id, BigInteger amount, BigInteger amounDiff)
     {
         EventManager.Instance.BankItemAdded(id, amount, amounDiff);
     }
 
-    public override void RaiseItemRemovedEvent(long id, int amount, int amounDiff)
+    public override void RaiseItemRemovedEvent(long id, BigInteger amount, BigInteger amounDiff)
     {
         EventManager.Instance.BankItemRemoved(id, amount, amounDiff);
     }
